app/requirements.txt
selenium
pylint
google-auth